requirements.txt
# Database Connectors
psycopg2-binary==2.9.9
cx_Oracle==8.3.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
sqlparse==0.4.4

# Testing
pytest==7.4.3
